# nest-template/openapi-schema.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/products": {
            "post": {
                "operationId": "ProductsController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProducts"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            },
            "get": {
                "operationId": "ProductsController_read",
                "parameters": [
                    {
                        "name": "",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/ReadProductsQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/products/{id}": {
            "put": {
                "operationId": "ProductsController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProducts"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "operationId": "ProductsController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "get": {
                "operationId": "ProductsController_readOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/ReadOneProductsQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/products/{id}/soft": {
            "delete": {
                "operationId": "ProductsController_softDelete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/order-_items": {
            "post": {
                "operationId": "Order_itemsController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrder_items"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            },
            "get": {
                "operationId": "Order_itemsController_read",
                "parameters": [
                    {
                        "name": "",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/ReadOrder_itemsQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/order-_items/{id}": {
            "put": {
                "operationId": "Order_itemsController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOrder_items"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "operationId": "Order_itemsController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "get": {
                "operationId": "Order_itemsController_readOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/ReadOneOrder_itemsQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/order-_items/{id}/soft": {
            "delete": {
                "operationId": "Order_itemsController_softDelete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/orders": {
            "post": {
                "operationId": "OrdersController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrders"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            },
            "get": {
                "operationId": "OrdersController_read",
                "parameters": [
                    {
                        "name": "",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/ReadOrdersQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/orders/{id}": {
            "put": {
                "operationId": "OrdersController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOrders"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "operationId": "OrdersController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "get": {
                "operationId": "OrdersController_readOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/ReadOneOrdersQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/orders/{id}/soft": {
            "delete": {
                "operationId": "OrdersController_softDelete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "info": {
        "title": "Nestjs Template API",
        "description": "",
        "version": "0.1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
            "CreateOrders": {
                "type": "object",
                "properties": {
                    "user_id": {
                        "nullable": true
                    },
                    "total_amount": {
                        "type": "number"
                    },
                    "placed_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "total_amount",
                    "placed_at"
                ],
                "additionalProperties": false
            },
            "ReadOneOrdersQuery": {
                "type": "object",
                "properties": {
                    "relations": {
                        "type": "object",
                        "properties": {
                            "user_id": {
                                "anyOf": [
                                    {
                                        "anyOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "boolean"
                                            }
                                        ],
                                        "type": "boolean"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ReadUsersRelations"
                                    }
                                ]
                            }
                        },
                        "required": [],
                        "additionalProperties": false
                    }
                },
                "required": [],
                "additionalProperties": false
            },
...
            "UpdateOrders": {
                "type": "object",
                "properties": {
                    "user_id": {
                        "nullable": true
                    },
                    "total_amount": {
                        "type": "number"
                    },
                    "placed_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [],
                "additionalProperties": false
            },
            "CreateOrder_items": {
                "type": "object",
                "properties": {
                    "order_id": {
                        "nullable": true
                    },
                    "product_id": {
                        "nullable": true
                    },
                    "quantity": {
                        "type": "number"
                    },
                    "price_at_purchase": {
                        "type": "number"
                    }
                },
                "required": [
                    "quantity",
                    "price_at_purchase"
                ],
                "additionalProperties": false
            },
            "ReadOneOrder_itemsQuery": {
                "type": "object",
                "properties": {
                    "relations": {
                        "type": "object",
                        "properties": {
                            "order_id": {
                                "anyOf": [
                                    {
                                        "anyOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "boolean"
                                            }
                                        ],
                                        "type": "boolean"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ReadOrdersRelations"
                                    }
                                ]
                            },
                            "product_id": {
                                "anyOf": [
                                    {
                                        "anyOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "boolean"
                                            }
                                        ],
                                        "type": "boolean"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ReadProductsRelations"
                                    }
                                ]
                            }
                        },
                        "required": [],
                        "additionalProperties": false
                    }
                },
                "required": [],
                "additionalProperties": false
            },
...
            "ReadOrder_itemsQuery": {
                "type": "object",
                "properties": {
                    "filters": {
                        "$ref": "#/components/schemas/ReadOrder_itemsFilters"
                    },
                    "relations": {
                        "type": "object",
                        "properties": {
                            "order_id": {
                                "anyOf": [
                                    {
                                        "anyOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "boolean"
                                            }
                                        ],
                                        "type": "boolean"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ReadOrdersRelations"
                                    }
                                ]
                            },
                            "product_id": {
                                "anyOf": [
                                    {
                                        "anyOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "boolean"
                                            }
                                        ],
                                        "type": "boolean"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ReadProductsRelations"
                                    }
                                ]
                            }
                        },
                        "required": [],
                        "additionalProperties": false
                    },
                    "orders": {
                        "$ref": "#/components/schemas/ReadOrder_itemsOrders"
                    },
                    "pagination": {
                        "type": "object",
                        "properties": {
                            "skip": {
                                "anyOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "number"
                                    }
                                ],
                                "type": "number",
                                "default": 0
                            },
                            "take": {
                                "anyOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "number"
                                    }
                                ],
                                "type": "number",
                                "default": 25
                            }
                        },
                        "required": [],
                        "additionalProperties": false,
                        "default": {
                            "skip": 0,
                            "take": 25
                        }
                    }
                },
                "required": [],
                "additionalProperties": false
            },
...
            "UpdateOrder_items": {
                "type": "object",
                "properties": {
                    "order_id": {
                        "nullable": true
                    },
                    "product_id": {
                        "nullable": true
                    },
                    "quantity": {
                        "type": "number"
                    },
                    "price_at_purchase": {
                        "type": "number"
                    }
                },
                "required": [],
                "additionalProperties": false
            },
            "CreateProducts": {
                "type": "object",
                "properties": {
                    "seller_id": {
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "type": "number"
                    },
                    "inventory_count": {
                        "type": "number"
                    }
                },
                "required": [
                    "name",
                    "price",
                    "inventory_count"
                ],
                "additionalProperties": false
            },
            "ReadOneProductsQuery": {
                "type": "object",
                "properties": {
                    "relations": {
                        "type": "object",
                        "properties": {
                            "seller_id": {
                                "anyOf": [
                                    {
                                        "anyOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "boolean"
                                            }
                                        ],
                                        "type": "boolean"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ReadSellersRelations"
                                    }
                                ]
                            }
                        },
                        "required": [],
                        "additionalProperties": false
                    }
                },
                "required": [],
                "additionalProperties": false
            },
...
            "ReadProductsQuery": {
                "type": "object",
                "properties": {
                    "filters": {
                        "$ref": "#/components/schemas/ReadProductsFilters"
                    },
                    "relations": {
                        "type": "object",
                        "properties": {
                            "seller_id": {
                                "anyOf": [
                                    {
                                        "anyOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "boolean"
                                            }
                                        ],
                                        "type": "boolean"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ReadSellersRelations"
                                    }
                                ]
                            }
                        },
                        "required": [],
                        "additionalProperties": false
                    },
                    "orders": {
                        "$ref": "#/components/schemas/ReadProductsOrders"
                    },
                    "pagination": {
                        "type": "object",
                        "properties": {
                            "skip": {
                                "anyOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "number"
                                    }
                                ],
                                "type": "number",
                                "default": 0
                            },
                            "take": {
                                "anyOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "number"
                                    }
                                ],
                                "type": "number",
                                "default": 25
                            }
                        },
                        "required": [],
                        "additionalProperties": false,
                        "default": {
                            "skip": 0,
                            "take": 25
                        }
                    }
                },
                "required": [],
                "additionalProperties": false
            },
...
            "UpdateProducts": {
                "type": "object",
                "properties": {
                    "seller_id": {
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "type": "number"
                    },
                    "inventory_count": {
                        "type": "number"
                    }
                },
                "required": [],
                "additionalProperties": false
            },
...
        }
    },
...
}